Add navigation configuration file and update imports
- Introduced `config.ts` to define `NavLink` interface and primary/secondary navigation links.
- Updated imports in `PrimaryNav` and `SecondaryNav` to reference the new `config.ts` instead of the removed `constants`.
- Deleted the old `types.ts` file as its content is now integrated into `config.ts`.
