--- src/components/TopNav/types.ts
@@ -1,6 +1,0 @@
-export interface NavLink {
-  href: string;
-  text: string;
-  children?: NavLink[];
-  id?: string;
-}
